Printing config as multi log line entity, makes it readable and grepable on startup
